+++ README.md
@@ -1,2 +1,2 @@
-## MEAN Docker
-sample app in mean stack
+## MEAN Project
+A sample project using Mongo, Express, Angular and Node
