+++ express-server/config/database.js
@@ -1,0 +1,4 @@
+module.exports = {
+    remoteUrl : process.env.DB_URL,
+    localUrl: 'mongodb://127.0.0.1:27017/mean-docker'
+};
