--- express-server/npm-debug.log
@@ -1,35 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ 'C:\\Program Files\\nodejs\\node.exe',
-1 verbose cli   'C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
-1 verbose cli   'start' ]
-2 info using npm@3.10.10
-3 info using node@v6.9.5
-4 verbose run-script [ 'prestart', 'start', 'poststart' ]
-5 info lifecycle express-server@0.0.0~prestart: express-server@0.0.0
-6 silly lifecycle express-server@0.0.0~prestart: no script for prestart, continuing
-7 info lifecycle express-server@0.0.0~start: express-server@0.0.0
-8 verbose lifecycle express-server@0.0.0~start: unsafe-perm in lifecycle true
-9 verbose lifecycle express-server@0.0.0~start: PATH: C:\Program Files\nodejs\node_modules\npm\bin\node-gyp-bin;D:\GitRepository\mean-docker\express-server\node_modules\.bin;C:\ProgramData\Oracle\Java\javapath;C:\windows\system32;C:\windows;C:\windows\System32\Wbem;C:\windows\System32\WindowsPowerShell\v1.0\;C:\Program Files (x86)\Windows Kits\8.1\Windows Performance Toolkit\;C:\Program Files\Microsoft SQL Server\110\Tools\Binn\;C:\Program Files (x86)\Microsoft SQL Server\110\Tools\Binn\;C:\Program Files\Microsoft SQL Server\110\DTS\Binn\;C:\Program Files (x86)\Microsoft SQL Server\110\Tools\Binn\ManagementStudio\;C:\Program Files (x86)\Microsoft Visual Studio 10.0\Common7\IDE\PrivateAssemblies\;C:\Program Files (x86)\Microsoft SQL Server\110\DTS\Binn\;C:\windows\system32\config\systemprofile\.dnx\bin;C:\Program Files\Microsoft DNX\Dnvm\;C:\Program Files\Microsoft SQL Server\120\Tools\Binn\;C:\Program Files\nodejs\;C:\Program Files\Git\cmd;C:\Program Files\Microsoft SQL Server\130\Tools\Binn\;D:\Program Files (x86)\Brackets\command;D:\Program Files\Ruby23-x64\bin;D:\Program Files\Git\cmd;D:\GitHub\DevKit\bin;C:\Program Files\Docker Toolbox;D:\Program Files\Heroku\bin;C:\Users\414122\AppData\Roaming\npm
-10 verbose lifecycle express-server@0.0.0~start: CWD: D:\GitRepository\mean-docker\express-server
-11 silly lifecycle express-server@0.0.0~start: Args: [ '/d /s /c', 'node server.js' ]
-12 verbose stack Error: kill ENOSYS
-12 verbose stack     at exports._errnoException (util.js:1022:11)
-12 verbose stack     at process.kill (internal/process.js:172:13)
-12 verbose stack     at process.listener (C:\Program Files\nodejs\node_modules\npm\node_modules\npmlog\node_modules\gauge\node_modules\signal-exit\index.js:86:15)
-12 verbose stack     at emitNone (events.js:91:20)
-12 verbose stack     at process.emit (events.js:185:7)
-12 verbose stack     at processEmit (C:\Program Files\nodejs\node_modules\npm\node_modules\npmlog\node_modules\gauge\node_modules\signal-exit\index.js:146:32)
-12 verbose stack     at processEmit (C:\Program Files\nodejs\node_modules\npm\node_modules\npm-registry-client\node_modules\npmlog\node_modules\gauge\node_modules\signal-exit\index.js:146:32)
-12 verbose stack     at Signal.wrap.onsignal (internal/process.js:199:44)
-13 verbose cwd D:\GitRepository\mean-docker\express-server
-14 error Windows_NT 6.1.7601
-15 error argv "C:\\Program Files\\nodejs\\node.exe" "C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "start"
-16 error node v6.9.5
-17 error npm  v3.10.10
-18 error code ENOSYS
-19 error errno ENOSYS
-20 error syscall kill
-21 error kill ENOSYS
-22 error If you need help, you may report this error at:
-22 error     <https://github.com/npm/npm/issues>
-23 verbose exit [ 1, true ]
