--- express-server/routes/api.js
@@ -1,62 +1,0 @@
-// Import dependencies
-const mongoose = require('mongoose');
-const express = require('express');
-const router = express.Router();
-
-// MongoDB URL 
-const dbHost = process.env.DB_URL || 'mongodb://127.0.0.1:27017/mean-docker';
-
-// Connect to mongodb
-mongoose.connect(dbHost);
-
-// create mongoose schema
-const userSchema = new mongoose.Schema({
-  name: String,
-  age: Number
-});
-
-// create mongoose model
-const User = mongoose.model('User', userSchema);
-
-/* GET api listing. */
-router.get('/', (req, res) => {
-		res.send('api works');
-});
-
-/* GET all users. */
-router.get('/users', (req, res) => {
-	User.find({}, (err, users) => {
-		if (err) res.status(500).send(error)
-
-		res.status(200).json(users);
-	});
-});
-
-/* GET one users. */
-router.get('/users/:id', (req, res) => {
-	User.findById(req.params.id, (err, users) => {
-		if (err) res.status(500).send(error)
-
-		res.status(200).json(users);
-	});
-});
-
-/* Create a user. */
-router.post('/users', (req, res) => {
-	let user = new User({
-		name: req.body.name,
-		age: req.body.age
-	});
-
-    console.log(user);
-    
-	user.save(error => {
-		if (error) res.status(500).send(error);
-
-		res.status(201).json({
-			message: 'User created successfully'
-		});
-	});
-});
-
-module.exports = router;
